wms-quantity: extracted reversibility of Split/Agg
The methods telling if Split and Aggregate are reversible
according to behaviour now naturally belong
to the wms-quantity Blok
